// schemas/user.ts
export const getUserInfoSchema = z.object({
  userId: z.string(),
  data: getUserInfoDataSchema,
});

export const getCurrentUserInfoSchema = getUserInfoDataSchema;
// searching
export const searchUsersDataSchema = z.object({
  count: z.boolean().optional(),
  limit: z
    .object({
      start: z.number(),
      end: z.number(),
    })
    .optional(),
  users: getUserInfoDataSchema,
});
export const searchUsersFiltersSchema = z.object({
  role: z.string().array().optional(),
  status: z.string().array().optional(),
  subscription: z.string().array().optional(),
});
export const searchUsersSchema = z.object({
  query: z.string().optional().default(""),
  filters: searchUsersFiltersSchema.optional(),
  data: searchUsersDataSchema,
});