// schemas/shared-schemas.ts
export const getUserSessionsDataSchema = z.object({
  count: z.optional(z.boolean()),
  limit: z.optional(
    z.object({
      start: z.number(),
      end: z.number(),
    })
  ),
  active: z.optional(z.boolean()),
});

export const getUserImagesDataSchema = z.object({
  count: z.boolean().optional(),
  limit: z
    .object({
      start: z.number(),
      end: z.number(),
    })
    .optional(),
  private: z.optional(z.boolean()),
});